+++ src/main/java/com/github/cao/awa/apricot/network/packet/factor/message/group/sent/GroupNormalMessageSentPacketFactor.java
@@ -20,7 +20,7 @@ public class GroupNormalMessageSentPacketFactor extends PacketFactor {
                 GroupMessageSender.create(request.getJSONObject("sender")),
                 request.getLong("group_id"),
                 request.getLong("time"),
-                request.getLong("message_id")
+                request.getInteger("message_id")
         );
     }
