+++ src/main/java/com/github/cao/awa/apricot/network/packet/receive/message/personal/received/PrivateFriendMessageReceivedPacket.java
@@ -12,9 +12,9 @@ public class PrivateFriendMessageReceivedPacket extends PrivateMessageReceivedPa
     private final MessageSender sender;
     private final long responseId;
     private final long timestamp;
-    private final long messageId;
+    private final int messageId;
 
-    public PrivateFriendMessageReceivedPacket(AssembledMessage message, long botId, MessageSender sender, long responseId, long timestamp, long messageId) {
+    public PrivateFriendMessageReceivedPacket(AssembledMessage message, long botId, MessageSender sender, long responseId, long timestamp, int messageId) {
         this.message = message;
         this.botId = botId;
         this.sender = sender;
@@ -47,7 +47,7 @@ public class PrivateFriendMessageReceivedPacket extends PrivateMessageReceivedPa
         return this.timestamp;
     }
 
-    public long getMessageId() {
+    public int getMessageId() {
         return this.messageId;
     }
